Fixed tests, did not account for regression module
