+++ tests/ignite/contrib/metrics/regression/test_manhattan_distance.py
@@ -1,44 +1,38 @@
 from ignite.exceptions import NotComputableError
-from ignite.contrib.metrics import MaximumAbsoluteError
+from ignite.contrib.metrics.regression import ManhattanDistance
 import torch
 import numpy as np
 import pytest
 
 
 def test_zero_div():
-    m = MaximumAbsoluteError()
+    m = ManhattanDistance()
     with pytest.raises(NotComputableError):
         m.compute()
 
 
-def test_maximum_absolute_error():
+def test_mahattan_distance():
     a = np.random.randn(4)
     b = np.random.randn(4)
     c = np.random.randn(4)
     d = np.random.randn(4)
     ground_truth = np.random.randn(4)
 
-    m = MaximumAbsoluteError()
+    m = ManhattanDistance()
     m.reset()
 
-    np_ans = -1
-
     m.update((torch.from_numpy(a), torch.from_numpy(ground_truth)))
-    np_max = np.max(np.abs((a - ground_truth)))
-    np_ans = np_max if np_max > np_ans else np_ans
+    np_ans = (a - ground_truth).sum()
     assert m.compute() == pytest.approx(np_ans)
 
     m.update((torch.from_numpy(b), torch.from_numpy(ground_truth)))
-    np_max = np.max(np.abs((b - ground_truth)))
-    np_ans = np_max if np_max > np_ans else np_ans
+    np_ans += (b - ground_truth).sum()
     assert m.compute() == pytest.approx(np_ans)
 
     m.update((torch.from_numpy(c), torch.from_numpy(ground_truth)))
-    np_max = np.max(np.abs((c - ground_truth)))
-    np_ans = np_max if np_max > np_ans else np_ans
+    np_ans += (c - ground_truth).sum()
     assert m.compute() == pytest.approx(np_ans)
 
     m.update((torch.from_numpy(d), torch.from_numpy(ground_truth)))
-    np_max = np.max(np.abs((d - ground_truth)))
-    np_ans = np_max if np_max > np_ans else np_ans
+    np_ans += (d - ground_truth).sum()
     assert m.compute() == pytest.approx(np_ans)
