+++ tests/ignite/contrib/metrics/regression/test_maximum_absolute_error.py
@@ -1,46 +1,44 @@
 from ignite.exceptions import NotComputableError
-from ignite.contrib.metrics import FractionalBias
+from ignite.contrib.metrics.regression import MaximumAbsoluteError
 import torch
 import numpy as np
 import pytest
 
 
 def test_zero_div():
-    m = FractionalBias()
+    m = MaximumAbsoluteError()
     with pytest.raises(NotComputableError):
         m.compute()
 
 
-def test_fractional_bias():
+def test_maximum_absolute_error():
     a = np.random.randn(4)
     b = np.random.randn(4)
     c = np.random.randn(4)
     d = np.random.randn(4)
     ground_truth = np.random.randn(4)
 
-    m = FractionalBias()
+    m = MaximumAbsoluteError()
     m.reset()
 
+    np_ans = -1
+
     m.update((torch.from_numpy(a), torch.from_numpy(ground_truth)))
-    np_sum = (2 * (a - ground_truth) / (a + ground_truth)).sum()
-    np_len = len(a)
-    np_ans = np_sum / np_len
+    np_max = np.max(np.abs((a - ground_truth)))
+    np_ans = np_max if np_max > np_ans else np_ans
     assert m.compute() == pytest.approx(np_ans)
 
     m.update((torch.from_numpy(b), torch.from_numpy(ground_truth)))
-    np_sum += (2 * (b - ground_truth) / (b + ground_truth)).sum()
-    np_len += len(b)
-    np_ans = np_sum / np_len
+    np_max = np.max(np.abs((b - ground_truth)))
+    np_ans = np_max if np_max > np_ans else np_ans
     assert m.compute() == pytest.approx(np_ans)
 
     m.update((torch.from_numpy(c), torch.from_numpy(ground_truth)))
-    np_sum += (2 * (c - ground_truth) / (c + ground_truth)).sum()
-    np_len += len(c)
-    np_ans = np_sum / np_len
+    np_max = np.max(np.abs((c - ground_truth)))
+    np_ans = np_max if np_max > np_ans else np_ans
     assert m.compute() == pytest.approx(np_ans)
 
     m.update((torch.from_numpy(d), torch.from_numpy(ground_truth)))
-    np_sum += (2 * (d - ground_truth) / (d + ground_truth)).sum()
-    np_len += len(d)
-    np_ans = np_sum / np_len
+    np_max = np.max(np.abs((d - ground_truth)))
+    np_ans = np_max if np_max > np_ans else np_ans
     assert m.compute() == pytest.approx(np_ans)
