+++ tests/ignite/contrib/metrics/regression/test_mean_error.py
@@ -1,25 +1,14 @@
 from ignite.exceptions import NotComputableError
-from ignite.contrib.metrics import MeanNormalizedBias
+from ignite.contrib.metrics.regression import MeanError
 import torch
 import numpy as np
 import pytest
 
 
 def test_zero_div():
-    m = MeanNormalizedBias()
+    m = MeanError()
     with pytest.raises(NotComputableError):
         m.compute()
-
-
-def test_zero_gt():
-    a = np.random.randn(4)
-    ground_truth = np.zeros(4)
-
-    m = MeanNormalizedBias()
-    m.reset()
-
-    with pytest.raises(NotComputableError):
-        m.update((torch.from_numpy(a), torch.from_numpy(ground_truth)))
 
 
 def test_mean_error():
@@ -29,29 +18,29 @@ def test_mean_error():
     d = np.random.randn(4)
     ground_truth = np.random.randn(4)
 
-    m = MeanNormalizedBias()
+    m = MeanError()
     m.reset()
 
     m.update((torch.from_numpy(a), torch.from_numpy(ground_truth)))
-    np_sum = ((a - ground_truth) / ground_truth).sum()
+    np_sum = (a - ground_truth).sum()
     np_len = len(a)
     np_ans = np_sum / np_len
     assert m.compute() == pytest.approx(np_ans)
 
     m.update((torch.from_numpy(b), torch.from_numpy(ground_truth)))
-    np_sum += ((b - ground_truth) / ground_truth).sum()
+    np_sum += (b - ground_truth).sum()
     np_len += len(b)
     np_ans = np_sum / np_len
     assert m.compute() == pytest.approx(np_ans)
 
     m.update((torch.from_numpy(c), torch.from_numpy(ground_truth)))
-    np_sum += ((c - ground_truth) / ground_truth).sum()
+    np_sum += (c - ground_truth).sum()
     np_len += len(c)
     np_ans = np_sum / np_len
     assert m.compute() == pytest.approx(np_ans)
 
     m.update((torch.from_numpy(d), torch.from_numpy(ground_truth)))
-    np_sum += ((d - ground_truth) / ground_truth).sum()
+    np_sum += (d - ground_truth).sum()
     np_len += len(d)
     np_ans = np_sum / np_len
     assert m.compute() == pytest.approx(np_ans)
